Removed utils folder and moved relevant code to specific folders
