+++ prive/datasets/dataset.py
@@ -7,7 +7,7 @@ import numpy as np
 import pandas as pd
 
 from prive.datasets.data_description import DataDescription
-from prive.utils.data import encode_data, index_split, get_dtype
+from .utils import encode_data, index_split, get_dtype
 
 # Helper function for parsing file-like objects
 def _parse_csv(fp, schema):
